+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/featured-image/index.tsx
@@ -3,14 +3,16 @@
  */
 import { useImageGenerator } from '@automattic/jetpack-ai-client';
 import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
-import { Button, Spinner } from '@wordpress/components';
+import { Button, Tooltip } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
-import { useCallback, useState } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
+import { useCallback, useRef, useState } from '@wordpress/element';
+import { __, sprintf } from '@wordpress/i18n';
+import { Icon, external } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
 import './style.scss';
+import UpgradePrompt from '../../../../blocks/ai-assistant/components/upgrade-prompt';
 import useAiFeature from '../../../../blocks/ai-assistant/hooks/use-ai-feature';
 import {
 	PLAN_TYPE_FREE,
@@ -21,6 +23,7 @@ import {
 import usePostContent from '../../hooks/use-post-content';
 import useSaveToMediaLibrary from '../../hooks/use-save-to-media-library';
 import AiAssistantModal from '../modal';
+import Carrousel, { CarrouselImageData, CarrouselImages } from './carrousel';
 import UsageCounter from './usage-counter';
 
 const FEATURED_IMAGE_FEATURE_NAME = 'featured-post-image';
@@ -31,19 +34,25 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 		useDispatch( 'core/edit-post' );
 	const { editPost, toggleEditorPanelOpened: toggleEditorPanelOpenedFromEditor } =
 		useDispatch( 'core/editor' );
+	const { clearSelectedBlock } = useDispatch( 'core/block-editor' );
+
+	const [ isFeaturedImageModalVisible, setIsFeaturedImageModalVisible ] = useState( false );
+	const [ images, setImages ] = useState< CarrouselImages >( [ { generating: true } ] );
+	const [ current, setCurrent ] = useState( 0 );
+	const pointer = useRef( 0 );
 
 	const { enableComplementaryArea } = useDispatch( 'core/interface' );
-	const [ isFeaturedImageModalVisible, setIsFeaturedImageModalVisible ] = useState( false );
-	const [ generating, setGenerating ] = useState( false );
-	const [ imageURL, setImageURL ] = useState( null );
-	const [ error, setError ] = useState( null );
 	const { generateImage } = useImageGenerator();
-	const { isLoading: isSavingToMediaLibrary, saveToMediaLibrary } = useSaveToMediaLibrary();
+	const { saveToMediaLibrary } = useSaveToMediaLibrary();
 	const { tracks } = useAnalytics();
 	const { recordEvent } = tracks;
 
+	const currentImage = images[ current ];
+	const currentPointer = images[ pointer.current ];
+
 	// Get feature data
 	const {
+		requireUpgrade,
 		requestsCount: allTimeRequestsCount,
 		requestsLimit: freeRequestsLimit,
 		usagePeriod,
@@ -82,12 +91,32 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 		increaseRequestsCount( featuredImageCost );
 	}, [ increaseRequestsCount, featuredImageCost ] );
 
+	/* Merge the image data with the new data. */
+	const updateImages = useCallback( ( data: CarrouselImageData, index ) => {
+		setImages( currentImages => {
+			const newImages = [ ...currentImages ];
+			newImages[ index ] = {
+				...newImages[ index ],
+				...data,
+			};
+			return newImages;
+		} );
+	}, [] );
+
+	const handlePreviousImage = useCallback( () => {
+		setCurrent( Math.max( current - 1, 0 ) );
+	}, [ current, setCurrent ] );
+
+	const handleNextImage = useCallback( () => {
+		setCurrent( Math.min( current + 1, images.length - 1 ) );
+	}, [ current, images.length ] );
+
 	/*
 	 * Function to generate a new image with the current value of the post content.
 	 */
 	const processImageGeneration = useCallback( () => {
-		setGenerating( true );
-		setError( null );
+		updateImages( { generating: true, error: null }, pointer.current );
+
 		generateImage( {
 			feature: FEATURED_IMAGE_FEATURE_NAME,
 			postContent,
@@ -96,17 +125,23 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 			.then( result => {
 				if ( result.data.length > 0 ) {
 					const image = 'data:image/png;base64,' + result.data[ 0 ].b64_json;
-					setImageURL( image );
+					updateImages( { image }, pointer.current );
 					updateRequestsCount();
+					saveToMediaLibrary( image )
+						.then( savedImage => {
+							updateImages( { libraryId: savedImage.id, generating: false }, pointer.current );
+							pointer.current += 1;
+						} )
+						.catch( () => {
+							updateImages( { generating: false }, pointer.current );
+							pointer.current += 1;
+						} );
 				}
 			} )
 			.catch( e => {
-				setError( e );
-			} )
-			.finally( () => {
-				setGenerating( false );
+				updateImages( { generating: false, error: e }, pointer.current );
 			} );
-	}, [ postContent, setGenerating, setImageURL, generateImage, updateRequestsCount ] );
+	}, [ updateImages, generateImage, postContent, updateRequestsCount, saveToMediaLibrary ] );
 
 	const toggleFeaturedImageModal = useCallback( () => {
 		setIsFeaturedImageModalVisible( ! isFeaturedImageModalVisible );
@@ -129,6 +164,7 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 		} );
 
 		processImageGeneration();
+		setCurrent( crrt => crrt + 1 );
 	}, [ processImageGeneration, recordEvent ] );
 
 	const handleTryAgain = useCallback( () => {
@@ -141,8 +177,10 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 	}, [ processImageGeneration, recordEvent ] );
 
 	const triggerComplementaryArea = useCallback( () => {
-		enableComplementaryArea( 'core/edit-post', 'edit-post/document' );
-	}, [ enableComplementaryArea ] );
+		// clear any block selection, because selected blocks have precedence on settings sidebar
+		clearSelectedBlock();
+		return enableComplementaryArea( 'core/edit-post', 'edit-post/document' );
+	}, [ clearSelectedBlock, enableComplementaryArea ] );
 
 	const handleAccept = useCallback( () => {
 		// track the accept/use image event
@@ -150,25 +188,41 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 			placement: JETPACK_SIDEBAR_PLACEMENT,
 		} );
 
-		saveToMediaLibrary( imageURL ).then( image => {
-			editPost( { featured_media: image.id } );
+		const setAsFeaturedImage = image => {
+			editPost( { featured_media: image } );
 			toggleFeaturedImageModal();
 
 			// Open the featured image panel for users to see the new image.
 			setTimeout( () => {
-				// If the panel is not opened, open it and then trigger the complementary area.
-				if ( ! isEditorPanelOpened( 'featured-image' ) ) {
-					toggleEditorPanelOpened?.( 'featured-image' ).then( () => {
-						triggerComplementaryArea();
-					} );
-				} else {
-					triggerComplementaryArea();
-				}
+				const isFeaturedImagePanelOpened = isEditorPanelOpened( 'featured-image' );
+				const isPostStatusPanelOpened = isEditorPanelOpened( 'post-status' );
+
+				// open the complementary area and then trigger the featured image panel.
+				triggerComplementaryArea().then( () => {
+					if ( ! isFeaturedImagePanelOpened ) {
+						toggleEditorPanelOpened( 'featured-image' );
+					}
+					// handle the case where the featured image panel is not present
+					if ( ! isPostStatusPanelOpened ) {
+						toggleEditorPanelOpened( 'post-status' );
+					}
+				} );
 			}, 500 );
-		} );
+		};
+
+		// If the image is already in the media library, use it directly, if it failed for some reason
+		// save it to the media library and then use it.
+		if ( images[ current ].libraryId ) {
+			setAsFeaturedImage( images[ current ].libraryId );
+		} else {
+			saveToMediaLibrary( images[ current ].image ).then( image => {
+				setAsFeaturedImage( image.id );
+			} );
+		}
 	}, [
+		current,
 		editPost,
-		imageURL,
+		images,
 		isEditorPanelOpened,
 		recordEvent,
 		saveToMediaLibrary,
@@ -177,17 +231,16 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 		triggerComplementaryArea,
 	] );
 
-	const modalTitleWhenGenerating = __( 'Generating featured image…', 'jetpack' );
-	const modalTitleWhenDone = __( 'Featured Image Generation', 'jetpack' );
+	const modalTitle = __( 'Generate a featured image with AI', 'jetpack' );
+	const costTooltipText = sprintf(
+		// Translators: %d is the cost of generating one image.
+		__( '%d requests per image', 'jetpack' ),
+		featuredImageCost
+	);
 
 	return (
 		<div>
-			<p>
-				{ __(
-					'Ask Jetpack AI to generate an image based on your post content, to use as the post featured image.',
-					'jetpack'
-				) }
-			</p>
+			<p>{ __( 'Create and use an AI generated featured image for your post.', 'jetpack' ) }</p>
 			<Button
 				onClick={ handleGenerate }
 				isBusy={ busy }
@@ -197,74 +250,83 @@ export default function FeaturedImage( { busy, disabled }: { busy: boolean; disa
 				{ __( 'Generate image', 'jetpack' ) }
 			</Button>
 			{ isFeaturedImageModalVisible && (
-				<AiAssistantModal
-					handleClose={ toggleFeaturedImageModal }
-					title={ generating ? modalTitleWhenGenerating : modalTitleWhenDone }
-				>
-					{ generating ? (
-						<div className="ai-assistant-featured-image__loading">
-							<Spinner
-								style={ {
-									width: '50px',
-									height: '50px',
-								} }
+				<AiAssistantModal handleClose={ toggleFeaturedImageModal } title={ modalTitle }>
+					<div className="ai-assistant-featured-image__content">
+						<div className="ai-assistant-featured-image__image-canvas">
+							{ ( requireUpgrade || notEnoughRequests ) && (
+								<UpgradePrompt
+									description={
+										notEnoughRequests
+											? sprintf(
+													// Translators: %d is the cost of generating a featured image.
+													__(
+														"Featured image generation costs %d requests per image. You don't have enough requests to generate another image.",
+														'jetpack'
+													),
+													featuredImageCost
+											  )
+											: null
+									}
+								/>
+							) }
+							<Carrousel
+								images={ images }
+								current={ current }
+								handlePreviousImage={ handlePreviousImage }
+								handleNextImage={ handleNextImage }
 							/>
 						</div>
-					) : (
-						<div className="ai-assistant-featured-image__content">
-							<div className="ai-assistant-featured-image__image-canvas">
-								{ error ? (
-									<div className="ai-assistant-featured-image__error">
-										{ __(
-											'An error occurred while generating the image. Please, try again!',
-											'jetpack'
-										) }
-										{ error?.message && (
-											<span className="ai-assistant-featured-image__error-message">
-												{ error?.message }
-											</span>
-										) }
-									</div>
-								) : (
-									<>
-										<img className="ai-assistant-featured-image__image" src={ imageURL } alt="" />
-										{ ! isUnlimited && featuredImageCost && requestsLimit && (
-											<UsageCounter
-												cost={ featuredImageCost }
-												currentLimit={ requestsLimit }
-												currentUsage={ requestsCount }
-											/>
-										) }
-									</>
+						<div className="ai-assistant-featured-image__actions">
+							<div className="ai-assistant-featured-image__actions-left">
+								{ ! isUnlimited && featuredImageCost && requestsLimit && (
+									<UsageCounter
+										cost={ featuredImageCost }
+										currentLimit={ requestsLimit }
+										currentUsage={ requestsCount }
+									/>
 								) }
 							</div>
-							<div className="ai-assistant-featured-image__actions">
-								{ ! error && (
+							<div className="ai-assistant-featured-image__actions-right">
+								<div className="ai-assistant-featured-image__action-buttons">
+									{ currentPointer?.error ? (
+										<Button onClick={ handleTryAgain } variant="secondary">
+											{ __( 'Try again', 'jetpack' ) }
+										</Button>
+									) : (
+										<Tooltip text={ costTooltipText } placement="bottom">
+											<Button
+												onClick={ handleRegenerate }
+												variant="secondary"
+												isBusy={ currentPointer?.generating }
+												disabled={ notEnoughRequests || currentPointer?.generating }
+											>
+												{ __( 'Generate again', 'jetpack' ) }
+											</Button>
+										</Tooltip>
+									) }
 									<Button
 										onClick={ handleAccept }
-										variant="secondary"
-										isBusy={ isSavingToMediaLibrary }
-										disabled={ isSavingToMediaLibrary }
+										variant="primary"
+										isBusy={ currentImage?.generating }
+										disabled={ ! currentImage?.image }
 									>
-										{ __( 'Save and use image', 'jetpack' ) }
+										{ __( 'Set as featured image', 'jetpack' ) }
 									</Button>
-								) }
-								{ error ? (
-									<Button onClick={ handleTryAgain } variant="secondary">
-										{ __( 'Try again', 'jetpack' ) }
-									</Button>
-								) : (
-									<Button
-										onClick={ handleRegenerate }
-										variant="secondary"
-										disabled={ isSavingToMediaLibrary || notEnoughRequests }
-									>
-										{ __( 'Generate another image', 'jetpack' ) }
-									</Button>
-								) }
+								</div>
 							</div>
 						</div>
-					) }
+					</div>
+					<div className="ai-assistant-featured-image__footer">
+						<Button
+							variant="link"
+							className="ai-assistant-featured-image__feedback-button"
+							href="https://jetpack.com/redirect/?source=jetpack-ai-feedback"
+							target="_blank"
+						>
+							<span>{ __( 'Provide feedback', 'jetpack' ) }</span>
+							<Icon icon={ external } className="icon" />
+						</Button>
+					</div>
 				</AiAssistantModal>
 			) }
 		</div>