+++ projects/plugins/jetpack/extensions/types.ts
@@ -20,7 +20,6 @@ export type SiteAIAssistantFeatureEndpointResponseProps = {
 	'site-require-upgrade': boolean;
 	'error-message'?: string;
 	'error-code'?: string;
-	'is-playground-visible'?: boolean;
 	'upgrade-type': UpgradeTypeProp;
 	'current-tier': TierProp;
 	'tier-plans': Array< TierProp >;